requirements.txt
msgpack-python>=0.4.8
pycrypto>=2.6.1
redis>=2.10.5
requests>=2.18.1
IPy==0.83
xmltodict==0.11.0